update delete coll route to check if a coll is owned by user

+++ server/routes/disccollections.ts
@@ -89,16 +89,23 @@ router.delete("/:collectionId", verifyUser, TryCatchAsync(async (req, res, next)
         return res.status(401).send("Unauthorized");
     }
     const { collectionId } = req.params
-    console.log("Someone tried to use API to post a disc collection");
-    console.log(`using the param ${collectionId}`)
-    const collectionToDelete = await DiscCollectionModel.findOneAndDelete(
-        {
-            _id: collectionId
-        }
-    )
-    console.log(`Collection ${collectionToDelete} is possibly removed from DB`);
-    await UserModel.findByIdAndUpdate(userId, { $pull: { collections: collectionId } });
-    res.status(200).json(collectionToDelete);
+    if (user.collections.includes(collectionId))
+    {
+        console.log("Someone tried to use API to post a disc collection");
+        console.log(`using the param ${collectionId}`)
+        const collectionToDelete = await DiscCollectionModel.findOneAndDelete(
+            {
+                _id: collectionId
+            }
+        )
+        console.log(`Collection ${collectionToDelete} is possibly removed from DB`);
+        await UserModel.findByIdAndUpdate(userId, { $pull: { collections: collectionId } });
+        res.status(200).json(collectionToDelete);
+    }
+    else
+    {
+        return res.status(401).send("Unauthorized! It's not yours!");
+    }
 }));
 
 module.exports = router;
